requirements.txt
alembic==1.4.0
apache-airflow==1.10.9
ExtractTable==2.2.0
numpy==1.18.1
pandas==0.25.3
paramiko==2.7.1
pendulum==1.4.4
prison==0.1.2
requests==2.22.0
SQLAlchemy==1.3.13
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.36.1
tweepy==3.10.0
typing==3.7.4.1
